feat: refine legacy app workflows
Add asset summaries, improve todo and daily mission handling, polish Bangumi scrolling, and harden PC file transfer plus backup encryption flows.

+++ app/src/main/Csharp/FileServer/FileInfo.cs
@@ -9,8 +9,8 @@ namespace FileServer
     public class FileInfo
     {
         public string fileName { get; set; }
-        public int fileSize { get; set; }
-        public FileInfo(string name,int size)
+        public long fileSize { get; set; }
+        public FileInfo(string name,long size)
         {
             fileName = name;
             fileSize = size;
